surface c code moved into its own file and header
