+++ tinywl/include/toplevel.h
@@ -47,4 +47,8 @@ void xdg_toplevel_request_move(struct wl_listener *listener, void *data);
 void xdg_toplevel_request_resize(struct wl_listener *listener, void *data);
 void xdg_toplevel_request_maximize(struct wl_listener *listener, void *data);
 void xdg_toplevel_request_fullscreen(struct wl_listener *listener, void *data);
+void xdg_toplevel_map(struct wl_listener *listener, void *data);
+void xdg_toplevel_unmap(struct wl_listener *listener, void *data);
+void xdg_toplevel_destroy(struct wl_listener *listener, void *data);
+void handle_toplevel_commit(struct wl_listener *listener, void *data);
 #endif
