+++ tinywl/tinywl.h
@@ -5,7 +5,6 @@
 #define WLR_USE_UNSTABLE
 #endif
 
-#include "wlr-layer-shell-unstable-v1-protocol.h"
 #include <stdbool.h>
 #include <stdint.h>
 #include <wayland-server-core.h>
@@ -72,6 +71,8 @@ struct tinywl_server *server_create(void);
 void server_destroy(struct tinywl_server *server);
 bool server_init(struct tinywl_server *server);
 const char *server_start(struct tinywl_server *server);
+
+void reset_cursor_mode(struct tinywl_server *server);
 
 void server_run(struct tinywl_server *server);
 void server_set_startup_command(const char *cmd);
